requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.0.4
django-allauth==0.62.1
django-bitfield==2.2.0
django-cors-headers==4.3.1
django-geo==0.8
django-js-asset==2.2.0
django-mptt==0.16.0
django-phonenumber-field==7.3.0
django-separatedvaluesfield==0.6.0
django-timezone-field==6.1.0
django-uuidfield==0.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-spectacular==0.27.2
geographiclib==2.0
geopy==2.4.1
idna==3.7
inflection==0.5.1
Js2Py==0.74
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lxml==5.2.1
packaging==24.0
phonenumbers==8.13.36
pillow==10.3.0
pipwin==0.5.2
psycopg2-2.9.9
pyjsparser==2.7.1
PyJWT==2.8.0
PyPrind==2.11.3
pySmartDL==1.3.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
six==1.16.0
soupsieve==2.5
South==1.0.2
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==2.2.1
